backend/requirements.txt
# Web framework
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20

# Database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
psycopg2-binary==2.9.10
alembic==1.14.1
pgvector==0.3.6

# Redis
redis==5.2.1

# Pydantic
pydantic==2.10.4
pydantic-settings==2.7.1

# AI / LLM
openai==1.58.1

# Google Cloud TTS
google-cloud-texttospeech==2.21.1

# Document processing
python-docx==1.1.2
openpyxl==3.1.5

# Embeddings / Vector
numpy==1.26.4

# Audio processing
soundfile==0.12.1
pydub==0.25.1

# Utilities
python-dotenv==1.0.1
httpx==0.28.1
jinja2==3.1.5
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# ASR (optional - needs GPU)
# transformers
# torch
# torchaudio
# onnxruntime
